requirements.txt
atlassian-python-api==4.0.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
Deprecated==1.2.18
idna==3.10
jmespath==1.0.1
oauthlib==3.2.2
python-dotenv==1.1.0
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
soupsieve==2.6
typing_extensions==4.13.0
urllib3==2.3.0
wrapt==1.17.2
